refactor: extract order summary
